feat: separate event management into @gland/events package
This commit introduces a significant restructuring of the event management system by separating it into its own package, `@gland/events`. Previously, event handling was tightly coupled within the core of the framework, making it more difficult to extend, maintain, and manage events across different protocols. By moving event-related logic into its own package, we gain several key benefits.
