--- packages/core/events/core/index.ts
@@ -1,3 +1,0 @@
-export * from './container';
-export * from './channel-proxy';
-export * from './queue';
