+++ packages/events/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@gland/http",
+  "name": "@gland/events",
   "version": "1.0.0",
   "author": "Mahdi",
   "license": "MIT",
@@ -11,10 +11,7 @@
   },
   "dependencies": {
     "@gland/common": "workspace:*",
-    "@gland/core": "workspace:*",
-    "@gland/events": "workspace:*",
-    "@medishn/toolkit": "^1.0.3",
-    "reflect-metadata": "^0.2.2"
+    "@medishn/toolkit": "^1.0.3"
   },
   "devDependencies": {
     "typescript": "^5.5.4"
