+++ packages/events/tsconfig.json
@@ -11,9 +11,6 @@
   "references": [
     {
       "path": "../common"
-    },
-    {
-      "path": "../events"
     }
   ]
 }
