+++ packages/http/tsconfig.json
@@ -14,6 +14,9 @@
     },
     {
       "path": "../core"
+    },
+    {
+      "path": "../events"
     }
   ]
 }
